Move and rename class.
