+++ s3/src/test/java/ch/cyberduck/core/s3/S3CredentialsConfiguratorTest.java
@@ -1,7 +1,7 @@
-package ch.cyberduck.core.sts;
+package ch.cyberduck.core.s3;
 
 /*
- * Copyright (c) 2002-2018 iterate GmbH. All rights reserved.
+ * Copyright (c) 2002-2023 iterate GmbH. All rights reserved.
  * https://cyberduck.io/
  *
  * This program is free software; you can redistribute it and/or modify
@@ -30,18 +30,18 @@ import java.io.File;
 
 import static org.junit.Assert.assertEquals;
 
-public class AWSProfileSTSCredentialsConfiguratorTest {
+public class S3CredentialsConfiguratorTest {
 
     @Test
     public void testConfigure() throws Exception {
-        new AWSProfileSTSCredentialsConfigurator(new DisabledX509TrustManager(), new DefaultX509KeyManager(), new DisabledPasswordCallback())
+        new S3CredentialsConfigurator(new DisabledX509TrustManager(), new DefaultX509KeyManager(), new DisabledPasswordCallback())
                 .reload().configure(new Host(new TestProtocol()));
     }
 
     @Test
     public void readFailureForInvalidAWSCredentialsProfileEntry() throws Exception {
         final Credentials credentials = new Credentials("test_s3_profile");
-        final Credentials verify = new AWSProfileSTSCredentialsConfigurator(LocalFactory.get(new File("src/test/resources/invalid/.aws").getAbsolutePath()),
+        final Credentials verify = new S3CredentialsConfigurator(LocalFactory.get(new File("src/test/resources/invalid/.aws").getAbsolutePath()),
                 new DisabledX509TrustManager(), new DefaultX509KeyManager(), new DisabledPasswordCallback())
                 .reload().configure(new Host(new TestProtocol(), StringUtils.EMPTY, credentials));
         assertEquals(credentials, verify);
@@ -49,7 +49,7 @@ public class AWSProfileSTSCredentialsConfiguratorTest {
 
     @Test
     public void readSuccessForValidAWSCredentialsProfileEntry() throws Exception {
-        final Credentials verify = new AWSProfileSTSCredentialsConfigurator(LocalFactory.get(new File("src/test/resources/valid/.aws").getAbsolutePath())
+        final Credentials verify = new S3CredentialsConfigurator(LocalFactory.get(new File("src/test/resources/valid/.aws").getAbsolutePath())
                 , new DisabledX509TrustManager(), new DefaultX509KeyManager(), new DisabledPasswordCallback())
                 .reload().configure(new Host(new TestProtocol(), StringUtils.EMPTY, new Credentials("test_s3_profile")));
         assertEquals("EXAMPLEKEYID", verify.getUsername());
